requirements.txt
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
flake8==5.0.4
flake8-broken-line==0.6.0
flake8-isort==6.0.0
flake8-plugin-utils==1.3.2
flake8-return==1.2.0
future==0.18.3
idna==3.4
importlib-metadata==4.2.0
isort==5.11.5
mccabe==0.7.0
numpy==1.21.6
opencv-python==4.7.0.72
pep8-naming==0.13.3
Pillow==9.4.0
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
pyTelegramBotAPI==4.10.0
python-dotenv==0.21.1
requests==2.28.2
typing_extensions==4.5.0
urllib3==1.26.14
zipp==3.15.0
